Split RichUpdateReport into internal
