+++ deltag.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+for t in `git tag`
+do
+  git tag -d $t
+done
+
