+++ librarymanagement/src/main/scala/sbt/internal/librarymanagement/syntax.scala
@@ -1,0 +1,10 @@
+package sbt
+package internal
+package librarymanagement
+
+import sbt.librarymanagement._
+
+object syntax {
+  implicit def richUpateReport(ur: UpdateReport): RichUpdateReport =
+    new RichUpdateReport(ur)
+}
